refactor(stella-tools): split the refresh tests out of engine/tests.rs
engine/tests.rs sits against the 1500-line ceiling and the new #3649 tests
pushed it over. Split into a sibling module rather than raising the baseline —
the baseline covers files that predate the guard, and a first-time crossing
gets split.
